Added PostOffice Skeleton code and change building constructor from using Vector2f to Vector2i.

+++ MailGame/src/Building.h
@@ -9,7 +9,7 @@ using namespace sf;
 class Building: public Entity
 {
 public:
-	Building(Vector2f pos, Direction dir, string spritePath);
+	Building(Vector2i pos, Direction dir, string spritePath);
 	virtual ~Building() {};
 
 	virtual void update(Time dt) = 0;
