+++ MailGame/src/House.h
@@ -9,7 +9,7 @@ using namespace sf;
 class House : public Building
 {
 public:
-	House(Vector2f pos, Direction dir);
+	House(Vector2i pos, Direction dir);
 	void update(Time dt);
 	//void render(RenderWindow * window);
